jupyter-requirements.txt
attrs==19.1.0
click==7.1.2
defusedxml==0.5.0
graphviz==0.16
hide-code==0.6.0
joblib==1.0.1
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.4.0
jupyter-core==4.6.0
jupyter-packaging==0.7.12
jupyter-telemetry==0.1.0
jupyterhub==1.3.0
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.11
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==5.6.1
nbformat==5.1.3
nbgrader==0.6.2
nbresuse==0.4.0
notebook==6.0.0
pandas==1.2.4
psutil==5.8.0
py4j==0.10.9.2
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4
qtconsole==5.2.0
setuptools==41.0.1
SQLAlchemy==1.3.8